+++ src/belpy/metashape/ms_utils.py
@@ -29,7 +29,7 @@ from typing import List
 from typing import Union
 from pathlib import Path
 
-from ..import_export.importing import read_opencv_calibration
+from ..io.importing import read_opencv_calibration
 
 """ License """
